+++ bonfire/modules/activities/config/config.php
@@ -1,7 +1,7 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
+<?php defined('BASEPATH') || exit('No direct script access allowed');
 
 $config['module_config'] = array(
-	'description'	=> 'Allows other modules to store user activity information.',
+	'description'   => 'Allows other modules to store user activity information.',
 	'author'		=> 'Bonfire Team',
-	'name'			=> 'Activities'
+	'name'			=> 'lang:bf_menu_activities',
 );