+++ bonfire/modules/emailer/config/config.php
@@ -1,10 +1,10 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
+<?php defined('BASEPATH') || exit('No direct script access allowed');
 
 $config['module_config'] = array(
-	'name'			=> 'Email Queue',
+	'name'			=> 'lang:bf_menu_emailer',
 	'description'	=> 'Queues emails to be sent in bursts throughout the day.',
 	'menus'	=> array(
-		'settings'	=> 'emailer/settings/menu'
+		'settings'	=> 'emailer/settings/menu',
 	),
-	'author'		=> 'Bonfire Team'
+	'author'		=> 'Bonfire Team',
 );