+++ bonfire/modules/emailer/controllers/settings.php
@@ -1,4 +1,4 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
+<?php defined('BASEPATH') || exit('No direct script access allowed');
 /**
  * Bonfire
  *
@@ -6,31 +6,25 @@
  *
  * @package   Bonfire
  * @author    Bonfire Dev Team
- * @copyright Copyright (c) 2011 - 2013, Bonfire Dev Team
- * @license   http://guides.cibonfire.com/license.html
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
  * @link      http://cibonfire.com
  * @since     Version 1.0
  * @filesource
  */
 
-// ------------------------------------------------------------------------
-
 /**
  * Emailer Settings Context
  *
  * Allows the management of the Emailer. Assists in setting up the proper email
- * settings, as well as editing the template and viewing emails that are in the queue.
+ * settings, as well as editing the template and viewing emails in the queue.
  *
- * @package    Bonfire
- * @subpackage Modules_Emailer
- * @category   Controllers
+ * @package    Bonfire\Modules\Emailer\Controllers\Settings
  * @author     Bonfire Dev Team
- * @link       http://guides.cibonfire.com/helpers/file_helpers.html
- *
+ * @link       http://cibonfire.com/docs/guides
  */
 class Settings extends Admin_Controller
 {
-
 	/**
 	 * Sets up the permissions and loads the language file
 	 *
@@ -43,18 +37,15 @@ class Settings extends Admin_Controller
 		$this->auth->restrict('Site.Settings.View');
 		$this->auth->restrict('Bonfire.Emailer.Manage');
 
-		Template::set_block('sub_nav', 'settings/_sub_nav');
-
 		$this->lang->load('emailer');
 
-	}//end __construct()
+		Assets::add_module_js('emailer', 'js/settings');
 
-	//--------------------------------------------------------------------
+		Template::set_block('sub_nav', 'settings/_sub_nav');
+	}
 
 	/**
 	 * Displays the emailer settings
-	 *
-	 * @access public
 	 *
 	 * @return void
 	 */
@@ -62,17 +53,14 @@ class Settings extends Admin_Controller
 	{
 		$this->load->library('form_validation');
 
-		if (isset($_POST['save']))
-		{
+		if (isset($_POST['save'])) {
 			$this->form_validation->set_rules('sender_email', 'lang:em_system_email', 'required|trim|valid_email|max_length[120]');
 			$this->form_validation->set_rules('protocol', 'lang:em_email_server', 'trim');
 
-			if ($this->input->post('protocol') == 'sendmail')
-			{
+            $protocol = $this->input->post('protocol');
+			if ($protocol == 'sendmail') {
 				$this->form_validation->set_rules('mailpath', 'lang:em_sendmail_path', 'required|trim');
-			}
-			elseif ($this->input->post('protocol') == 'smtp')
-			{
+			} elseif ($protocol == 'smtp') {
 				$this->form_validation->set_rules('smtp_host', 'lang:em_smtp_address', 'required|trim');
 				$this->form_validation->set_rules('smtp_user', 'lang:em_smtp_username', 'trim');
 				$this->form_validation->set_rules('smtp_pass', 'lang:em_smtp_password', 'trim|matches_pattern[[A-Za-z0-9!@#\%$^&+=]{2,20}]');
@@ -80,94 +68,72 @@ class Settings extends Admin_Controller
 				$this->form_validation->set_rules('smtp_timeout', 'lang:em_smtp_timeout', 'trim|numeric');
 			}
 
-			if ($this->form_validation->run() !== FALSE)
-			{
+			if ($this->form_validation->run() !== false) {
 				$data = array(
-						array('name' => 'sender_email', 'value' => $this->input->post('sender_email')),
-						array('name' => 'mailtype', 'value' => $this->input->post('mailtype')),
-						array('name' => 'protocol', 'value' => $_POST['protocol']),
-						array('name' => 'mailpath', 'value' => $_POST['mailpath']),
-						array('name' => 'smtp_host', 'value' => isset($_POST['smtp_host']) ? $_POST['smtp_host'] : ''),
-						array('name' => 'smtp_user', 'value' => isset($_POST['smtp_user']) ? $_POST['smtp_user'] : ''),
-						array('name' => 'smtp_pass', 'value' => isset($_POST['smtp_pass']) ? $_POST['smtp_pass'] : ''),
-						array('name' => 'smtp_port', 'value' => isset($_POST['smtp_port']) ? $_POST['smtp_port'] : ''),
-						array('name' => 'smtp_timeout', 'value' => isset($_POST['smtp_timeout']) ? $_POST['smtp_timeout'] : '5')
-					 );
+                    array('name' => 'sender_email', 'value' => $this->input->post('sender_email')),
+                    array('name' => 'mailtype',     'value' => $this->input->post('mailtype')),
+                    array('name' => 'protocol',     'value' => $protocol),
+                    array('name' => 'mailpath',     'value' => $_POST['mailpath']),
+                    array('name' => 'smtp_host',    'value' => isset($_POST['smtp_host']) ? $_POST['smtp_host'] : ''),
+                    array('name' => 'smtp_user',    'value' => isset($_POST['smtp_user']) ? $_POST['smtp_user'] : ''),
+                    array('name' => 'smtp_pass',    'value' => isset($_POST['smtp_pass']) ? $_POST['smtp_pass'] : ''),
+                    array('name' => 'smtp_port',    'value' => isset($_POST['smtp_port']) ? $_POST['smtp_port'] : ''),
+                    array('name' => 'smtp_timeout', 'value' => isset($_POST['smtp_timeout']) ? $_POST['smtp_timeout'] : '5'),
+                );
 
-				$updated = FALSE;
-				// save the settings to the db
+				// Save the settings to the db
 				$updated = $this->settings_model->update_batch($data, 'name');
 
-				if ($updated)
-				{
-					// Success, so reload the page, so they can see their settings
+				if ($updated) {
+					// Success, reload the page so they can see their settings
 					Template::set_message('Email settings successfully saved.', 'success');
-					redirect(SITE_AREA .'/settings/emailer');
-				}
-				else
-				{
+					redirect(SITE_AREA . '/settings/emailer');
+				} else {
 					Template::set_message('There was an error saving your settings.', 'error');
 				}
-			}
-			else
-			{
+			} else {
 				Template::set_message('There was an error saving your settings.', 'error');
 			}
-		}//end if
+		}
 
 		// Load our current settings
-		$settings = $this->settings_model->select('name,value')->find_all_by('module', 'email');
+		$settings = $this->settings_model->select('name,value')
+                                         ->find_all_by('module', 'email');
+
 		Template::set($settings);
-
-		Assets::add_module_js('emailer', 'js/settings');
-
 		Template::set('toolbar_title', 'Email Settings');
 
 		Template::render();
-
-	}//end index()
-
-	//--------------------------------------------------------------------
+	}
 
 	/**
 	 * Display the screen to edit the email templagtes
-	 *
-	 * @access public
 	 *
 	 * @return void
 	 */
 	public function template()
 	{
-		if (isset($_POST['save']))
-		{
+		if (isset($_POST['save'])) {
 			$header = $_POST['header'];
 			$footer = $_POST['footer'];
 
 			$this->load->helper('file');
 
-			write_file(BFPATH .'modules/emailer/views/email/_header.php', $header, 'w+');
-			write_file(BFPATH .'modules/emailer/views/email/_footer.php', $footer, 'w+');
+			write_file(BFPATH . 'modules/emailer/views/email/_header.php', $header, 'w+');
+			write_file(BFPATH . 'modules/emailer/views/email/_footer.php', $footer, 'w+');
 
 			Template::set_message('Template successfully saved.', 'success');
 
-			redirect(SITE_AREA .'/settings/emailer/template');
+			redirect(SITE_AREA . '/settings/emailer/template');
 		}
-
-
-		Assets::add_js(Template::theme_url('js/editors/ace/ace.js'));
-		Assets::add_js(Template::theme_url('js/editors/ace/theme-monokai.js'));
-		Assets::add_js(Template::theme_url('js/editors/ace/mode-html.js'));
-		Assets::add_module_js('emailer', 'js/ace');
 
 		Template::set('toolbar_title', lang('em_email_template'));
 
 		Template::render();
-	}//end template()
-
-	//--------------------------------------------------------------------
+	}
 
 	/**
-	 * @access puublic
+	 * @todo Remove this?
 	 *
 	 * @return void
 	 */
@@ -175,111 +141,87 @@ class Settings extends Admin_Controller
 	{
 		Template::set('toolbar_title', lang('em_email_contents'));
 		Template::render();
-
-	}//end emails()
-
-	//--------------------------------------------------------------------
+	}
 
 	/**
 	 * Send a test email
-	 *
-	 * @access public
 	 *
 	 * @return void
 	 */
 	public function test()
 	{
-		if (!isset($_POST['test']))
-		{
+		if ( ! isset($_POST['test'])) {
 			$this->security->csrf_show_error();
 		}
 
 		$this->load->library('emailer');
-		$this->emailer->enable_debug(TRUE);
+		$this->emailer->enable_debug(true);
 
 		$data = array(
-				'to'		=> $this->input->post('email'),
-				'subject'	=> lang('em_test_mail_subject'),
-				'message'	=> lang('em_test_mail_body')
-			 );
+            'to'		=> $this->input->post('email'),
+            'subject'	=> lang('em_test_mail_subject'),
+            'message'	=> lang('em_test_mail_body'),
+         );
 
-		$success = $this->emailer->send($data, FALSE);
+		$success = $this->emailer->send($data, false);
 
 		Template::set('success', $success);
 		Template::set('debug', $this->emailer->debug_message);
+
 		Template::render();
-
-	}//end test()
-
-	//--------------------------------------------------------------------
+	}
 
 	/**
-	 * Displays all of the emails currently in the queue to be sent.
-	 *
-	 * @access public
+	 * Display all of the emails currently in the queue.
 	 *
 	 * @return void
 	 */
 	public function queue()
 	{
-		$this->load->library('pagination');
-
 		$offset = $this->uri->segment(5);
 
-		$this->load->model('Emailer_model', 'emailer_model', TRUE);
+		$this->load->library('pagination');
+		$this->load->model('Emailer_model', 'emailer_model', true);
 
 		// Deleting anything?
-		if (isset($_POST['delete']))
-		{
+		if (isset($_POST['delete'])) {
 			$checked = $this->input->post('checked');
-			if (is_array($checked) && count($checked))
-			{
-				$result = FALSE;
-				foreach ($checked as $pid)
-				{
+			if (is_array($checked) && count($checked)) {
+				$result = false;
+				foreach ($checked as $pid) {
 					$result = $this->emailer_model->delete($pid);
 				}
 
-				if ($result)
-				{
+				if ($result) {
 					Template::set_message(sprintf(lang('em_delete_success'), count($checked)), 'success');
-				}
-				else
-				{
+				} else {
 					Template::set_message(sprintf(lang('em_delete_failure'), $this->emailer_model->error), 'error');
 				}
-			}
-			else
-			{
+			} else {
 				Template::set_message(sprintf(lang('em_delete_error'), $this->emailer_model->error), 'error');
 			}
-		}
-		elseif (isset($_POST['force_process']))
-		{
+		} elseif (isset($_POST['force_process'])) {
 			$this->load->library('emailer');
-			$this->emailer->enable_debug(TRUE);
+			$this->emailer->enable_debug(true);
 
 			// Use ob to catch output designed for CRON only
 			ob_start();
 			$success = $this->emailer->process_queue();
 			ob_end_clean();
 
-			if ( ! $success)
-			{
+			if ( ! $success) {
 				Template::set('email_debug', $this->emailer->debug_message);
 			}
-		}
-		elseif (isset($_POST['insert_test']))
-		{
+		} elseif (isset($_POST['insert_test'])) {
 			$this->load->library('emailer');
 
 			$data = array(
 				'to'		=> $this->settings_lib->item('site.system_email'),
 				'subject'	=> lang('em_test_mail_subject'),
-				'message'	=> lang('em_test_mail_body')
+				'message'	=> lang('em_test_mail_body'),
 			);
 
-			$this->emailer->send($data, TRUE);
+			$this->emailer->send($data, true);
 		}
 
 		Template::set('emails', $this->emailer_model->limit($this->limit, $offset)->find_all());
@@ -288,125 +230,102 @@ class Settings extends Admin_Controller
 
 		$total_emails = $this->emailer_model->count_all();
 
-		$this->pager['base_url'] = site_url(SITE_AREA .'/settings/emailer/queue');
-		$this->pager['total_rows'] = $total_emails;
-		$this->pager['per_page'] = $this->limit;
+		$this->pager['base_url']    = site_url(SITE_AREA . '/settings/emailer/queue');
+		$this->pager['total_rows']  = $total_emails;
+		$this->pager['per_page']    = $this->limit;
 		$this->pager['uri_segment']	= 5;
 
 		$this->pagination->initialize($this->pager);
 
 		Template::set('toolbar_title', lang('em_emailer_queue'));
+
 		Template::render();
-
-	}//end queue()
-
-	//--------------------------------------------------------------------
+	}
 
 	/**
 	 * Displays a preview of the email as stored in the database.
-	 *
-	 * @access public
 	 *
 	 * @param int $id An INT with the ID of the email to preview from the queue.
 	 *
 	 * @return void
 	 */
-	public function preview($id=0)
+	public function preview($id = 0)
 	{
-		$this->output->enable_profiler(FALSE);
-
+		$this->output->enable_profiler(false);
 		$this->load->model('emailer/emailer_model');
 
-		if (!empty($id) && is_numeric($id))
-		{
+		if ( ! empty($id) && is_numeric($id)) {
 			$email = $this->emailer_model->find($id);
 
-			if ($email)
-			{
+			if ($email) {
 				Template::set('email', $email);
-
 				Template::render('blank');
 			}
 		}
-
-	}//end preview()
+	}
 
 	/**
 	 * Create a new email and send to selected recipents
-	 *
-	 * @access public
 	 *
 	 * @return void
 	 */
 	public function create()
 	{
-
 		$this->load->model('users/user_model');
 		$this->load->library('emailer');
 
-		if (isset($_POST['create']))
-		{
-			// validate subject, content and recipients
+		if (isset($_POST['create'])) {
+			// Validate subject, content and recipients
 			$this->form_validation->set_rules('email_subject', 'lang:em_email_subject', 'required|trim|min_length[1]|max_length[255]');
 			$this->form_validation->set_rules('email_content', 'lang:em_email_content', 'required|trim|min_length[1]');
 			$this->form_validation->set_rules('checked', 'lang:bf_users', 'required');
 
-			if ($this->form_validation->run() === FALSE)
-			{
+			if ($this->form_validation->run() === false) {
+                // @todo This shouldn't be necessary, is set_value() not used in the view?
 				Template::set('email_subject', $this->input->post('email_subject'));
 				Template::set('email_content', $this->input->post('email_content'));
 				Template::set('checked', $this->input->post('checked'));
-			}
-			else
-			{
+			} else {
 				$data = array (
 					'subject'	=> $this->input->post('email_subject'),
 					'message'	=> $this->input->post('email_content'),
 				);
-
 				$checked = $this->input->post('checked');
+
 				$success_count = 0;
-				if (is_array($checked) && count($checked))
-				{
-					$result = FALSE;
-					foreach ($checked as $user_id)
-					{
-						//get the email from user_id
+				if (is_array($checked) && count($checked)) {
+					$result = false;
+					foreach ($checked as $user_id) {
+						// Get the email from $user_id
 						$user = $this->user_model->find($user_id);
-						if ($user != NULL){
+						if ($user != null) {
 							$data['to'] = $user->email;
-							$result = $this->emailer->send($data,TRUE);
-							if ($result) $success_count++;
+
+							$result = $this->emailer->send($data, true);
+							if ($result) {
+                                $success_count++;
+                            }
 						}
-
 					}
 
-					if ($result)
-					{
-						Template::set_message($success_count .' '. lang('em_create_email_success'), 'success');
+					if ($result) {
+						Template::set_message($success_count . ' ' . lang('em_create_email_success'), 'success');
 						redirect(SITE_AREA . '/settings/emailer/queue');
-					}
-					else
-					{
+					} else {
 						Template::set_message(sprintf(lang('em_create_email_failure'), $this->user_model->error), 'error');
 					}
-				}
-				else
-				{
+				} else {
 					Template::set_message(sprintf(lang('em_create_email_error'), $this->user_model->error), 'error');
-				}//end if
-			}//end if
-		}//end if
+				}
+			}
+		}//end if (isset($_POST['create']))
 
 		$users = $this->user_model->where('users.deleted', 0)->find_all();
+
 		Template::set('users', $users);
 		Template::set('toolbar_title', lang('em_create_email'));
-		Template::render();
 
-	}//end create()
-
-	//--------------------------------------------------------------------
-}//end class
-
-// End Admin class
+        Template::render();
+	}
+}
 /* End of file settings.php */