+++ bonfire/modules/permissions/config/config.php
@@ -1,6 +1,7 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
+<?php defined('BASEPATH') || exit('No direct script access allowed');
 
 $config['module_config'] = array(
 	'description'	=> 'Manages the Permissions available to the Roles.',
-	'author'		=> 'Bonfire Team'
+	'author'		=> 'Bonfire Team',
+    'name'          => 'lang:bf_menu_permissions',
 );