+++ bonfire/modules/users/config/config.php
@@ -1,9 +1,10 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
+<?php defined('BASEPATH') || exit('No direct script access allowed');
 
 $config['module_config'] = array(
 	'description'	=> 'Allows users to exist in Bonfire.',
 	'author'		=> 'Bonfire Team',
 	'weights'		=> array(
-		'settings'	=> 1
-	)
+		'settings'	=> 1,
+	),
+    'name'          => 'lang:bf_menu_users',
 );