+++ bonfire/docs/changelog.md
@@ -19,11 +19,13 @@ Additional Changes:
 * Many improvements to the Brazilian Portugeuse language handling.
 * Improved support for translations in the Builder.
 
+
 New Features:
 * Images controller can reference images from a module using the module=... parameter.
 * Template class now provides a `check_segment` method in addition to check_class and check_method.
 * Brand new documentation system that allows splitting your user and dev-related docs, and searching docs.
 * CSRF protection can be bypassed for individual controllers by setting `csrf_ignored_controllers` setting in the site's main index file.
+* Now works with [Sparks](http://getsparks.org) out of the box.
 
 
 Closes Issues: